logging utilities, backend.h, backend/session.h added
Decided to use a package per file approach for this. This may become
messy once more of the library is implemented, but for now it makes
sense. The mess is mostly due to the fact that cffi-grovel doesn't work
with asdf's package inffered system. Here are the reasons I stuck with it:
 + Helper functions can be hidden from different parts of the system
 + We can make the dependencies between the files more explicit
 + Users can keep a large amount of symbols from being exported