Add wlroots configuration information
Add the various configuration definitions to cl:*features if they are
present in wlroots. This requires new cffi-grovel functionality that
was just added.

The rationale for adding to *features* is that it makes it very easy
to conditionally compile code to avoid undefined foreign function
warnings.

The downside is that it requires recompilation if wlroots is updated
and these features are added/removed.